requirements.txt
fastapi
uvicorn
pydantic
pymongo
langchain
langchain-core
langchain-google-genai
python-dotenv
numpy
pandas
